refactor: imrpovement of factory

+++ Source/Infrastructure/Database/FactoryDatabase.ts
@@ -9,6 +9,8 @@ import {
     type IMSSQLDatabaseOptions,
     type IPostgresDatabaseOptions
 } from '@/Infrastructure/Database/Creator/index.js';
+import { AndesiteError } from '@/Common/Error/index.js';
+import { InfrastructureDatabaseKeys } from '@/Common/Error/Enum/index.js';
 
 /**
  * FactoryDatabase class.
@@ -41,12 +43,20 @@ export class FactoryDatabase {
      * @param name - The name of the database
      * @param type - The type of the database (ex: postgres, better-sqlite, mssql)
      * @param options - The options of the database. ({@link IPostgresDatabaseOptions} or {@link IBetterSQLiteDatabaseOptions} or {@link IMSSQLDatabaseOptions})
+     *
+     * @throws ({@link AndesiteError}) - If the database is already registered with the same name. ({@link InfrastructureDatabaseKeys.DATABASE_ALREADY_REGISTERED})
+     * @throws ({@link AndesiteError}) - If the database is not connected ({@link InfrastructureDatabaseKeys.DATABASE_NOT_CONNECTED})
      */
-    public register(
+    public async register(
         name: string,
         type: 'postgres' | 'better-sqlite' | 'mssql',
         options: IPostgresDatabaseOptions | IBetterSQLiteDatabaseOptions | IMSSQLDatabaseOptions
-    ): void {
+    ): Promise<void> {
+        if (this._database.has(name))
+            throw new AndesiteError({
+                messageKey: InfrastructureDatabaseKeys.DATABASE_ALREADY_REGISTERED,
+                detail: { name }
+            });
         let creator: AbstractCreator | undefined = undefined;
         if (type === 'postgres')
             creator = new PostgresCreator(options as IPostgresDatabaseOptions);
@@ -54,18 +64,27 @@ export class FactoryDatabase {
             creator = new BetterSQLiteCreator(options as IBetterSQLiteDatabaseOptions);
         else if (type === 'mssql')
             creator = new MSSQLCreator(options as IMSSQLDatabaseOptions);
-        if (creator)
+        if (creator) {
+            await creator.connection();
             this._database.set(name, creator);
+        }
     }
 
     /**
      * Unregister a database by name.
      *
      * @param name - The name of the database to unregister
+     *
+     * @throws ({@link AndesiteError}) - If the database is not registered with the same name. ({@link InfrastructureDatabaseKeys.DATABASE_ALREADY_NOT_REGISTERED})
      */
     public async unregister(name: string): Promise<void> {
+        if (!this._database.has(name))
+            throw new AndesiteError({
+                messageKey: InfrastructureDatabaseKeys.DATABASE_ALREADY_NOT_REGISTERED,
+                detail: { name }
+            });
         const database: AbstractCreator = this._database.get(name) as AbstractCreator;
-        if (database.isConnected())
+        if (await database.isConnected())
             await database.disconnection();
         this._database.delete(name);
     }
@@ -75,12 +94,17 @@ export class FactoryDatabase {
      *
      * @param name - The name of the database to get
      *
+     * @throws ({@link AndesiteError}) - If the database is not registered with the same name. ({@link InfrastructureDatabaseKeys.DATABASE_NOT_REGISTERED})
+     *
      * @returns The ({@link Knex}) instance
      */
-    public get(name: string): Knex | undefined {
+    public get(name: string): Knex {
+        if (!this._database.has(name))
+            throw new AndesiteError({
+                messageKey: InfrastructureDatabaseKeys.DATABASE_NOT_REGISTERED,
+                detail: { name }
+            });
         const database: AbstractCreator = this._database.get(name) as AbstractCreator;
-        if (!database.isConnected())
-            database.connection();
         return database.database;
     }
 }
